requirements.txt
# Core compute libs
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3

# NLP 
spacy>=3.7
nltk>=3.9

# DL 
torch>=2.6
transformers>=4.56

# REST Frameworks
fastapi==0.115.12
uvicorn==0.34

# Notebooks for viz & docs
jupyterlab>=4.0
ipykernel>=6.30

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock>=3.14.0
pytest-asyncio==0.21.1

# Linters and formatters
ruff>=0.12.0
black>=23.0.0
black[jupyter]>=23.0.0

# Utilities
python-dotenv==1.1.0
tqdm==4.67.1
wandb>=0.21
matplotlib>=3.8
chardet>=5.2
wordcloud==1.9.4